requirements.txt
# SIMAFI Web — dependencias para desarrollo y despliegue (VPS Linux + Gunicorn)
# Instalación: pip install -r requirements.txt

Django>=4.2,<5
PyMySQL>=1.1
cryptography>=45.0
djangorestframework>=3.14
django-widget-tweaks>=1.5
python-dotenv>=1.0
gunicorn>=22.0

# Exportaciones e informes (administrativo, configuración)
openpyxl>=3.1
reportlab>=4.0
Pillow>=10.0

# Opcional según módulos activos (descomentar si hace falta)
# pyproj>=3.6
# mysqlclient>=2.2
